Update PHPCS Configuration
Rename the ruleset for the projects PHP CodeSniffer rules from "phpcs.xml" to "ruleset.xml" (which is the official name required for the project to become it's own custom ruleset).
We stop using Nosco Systems' custom ruleset, in favour of the now industry-standard PSR2 FIG recommendations, and exclude some of Squiz's rules on commenting that Object Calisthenics include, because we don't like them.